requirements.txt
# Unified requirements for Luna Home Agent

# Core web and configuration
fastapi>=0.111.1
uvicorn>=0.30.5
pydantic>=2.8.2
httpx>=0.27.2
requests>=2.32.0
PyYAML>=6.0.2
loguru>=0.7.2

# GUI frameworks
PyQt5>=5.15.10
PySide6>=6.7.1

# OCR utilities
pillow>=10.0.0
mss>=9.0.1
keyboard>=0.13.5
paddlepaddle>=2.5.1
paddleocr>=2.7.0

# JSON schema validation
jsonschema>=4.21.1

# STT and audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
soundcard>=0.4.2
numpy>=1.24.0
scipy>=1.10.0
webrtcvad>=2.0.10
faster-whisper>=1.0.1
librosa>=0.10.0
scikit-learn>=1.3.0
pycaw>=20231210
comtypes>=1.2.0

# Web search helpers
duckduckgo-search>=5.3.0
beautifulsoup4>=4.12.3
lxml>=5.2.2

# Optional Windows notifications
win10toast>=0.9; platform_system == "Windows"
winotify>=1.1.0; platform_system == "Windows"

